psybot/requirements.txt
python-dotenv
SQLAlchemy
pymysql
oracledb
pymongo
pandas
datasets
fastapi
uvicorn
transformers
torch
sentencepiece
accelerate
